requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
python-decouple==3.8
requests==2.31.0
python-telegram-bot==20.7
google-generativeai==0.3.2
pillow
django-allauth